Add fix increament tick on continuous scale, transpose on bar chart and new rank function on Sorted barline plot  [v3.2.2]
[PCR #2016062801]

1. 對連續型座標軸提供固定間距 Tick (因為是非原生元件，需額外寫在繪圖程序裡面，故目前只提供給 Y 軸，主座標軸)
2. 在 MtbGraph.BarChart.Chart 元件中，提供 Transpose 的功能
3. 在 Sorted Bar-Line plot 提供只針對正數排序的功能

+++ Component/Scale/Adapter_CateTick.cs
@@ -30,7 +30,7 @@ namespace MtbGraph.Component.Scale
         {
             get
             {
-                return _tick.Increament;
+                return (int)_tick.Increament;
             }
             set
             {
